dump/requirements.txt
certifi
charset-normalizer
colorama
customtkinter
darkdetect
diffusers
filelock
huggingface-hub
idna==3.4
importlib-metadata
numpy
packaging
Pillow==9.2.0
pyparsing
PyYAML
regex==2022.9.13
requests
tk==0.1.0
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
urllib3
zipp